requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.6
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
idna==3.10
keras==3.7.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
protobuf==5.29.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
wrapt==1.17.0
